requirements.txt
Django==2.1.4
djangorestframework==3.9.0
Pillow==5.3.0
pkg-resources==0.0.0
python-dotenv==0.10.1
pytz==2018.7
